refactor: Added helper module and moved private class methods to the helper
