+++ lib/string_calculator/helpers.rb
@@ -1,0 +1,31 @@
+# lib/string_calculator/helpers.rb
+# frozen_string_literal: true
+
+# Helper methods for the StringCalculator pipeline
+module StringCalculatorHelpers
+  # Helper methods for the Calculator class
+
+  def validate_input(input)
+    # To do
+  end
+
+  def check_for_negatives(numbers)
+    # To do
+  end
+
+  def parse_components(input)
+    # To do
+  end
+
+  def parse_numbers(input)
+    # To do
+  end
+
+  def parse_delimiters(input)
+    # To do
+  end
+
+  def add_numbers(numbers)
+    # To do
+  end
+end
